pyproject.toml
[project]
name = "adk-test-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10.19,<=3.10.19"
dependencies = [
    "google-adk (>=1.19.0,<2.0.0)",
    "google-genai (>=1.52.0,<2.0.0)"
]

[tool.poetry]
package-mode=false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "mcp (>=1.22.0,<2.0.0)"
]
